Move one of the windows examples over to common src
